+++ index.js
@@ -86,14 +86,6 @@ export { default as Link } from './packages/link';
 
 export { default as Modal } from './packages/modal';
 
-export {
-  SubFooter,
-  ContentBlock,
-  NavBlock,
-  ButtonBlock,
-  SocialBlock,
-} from './packages/subfooter';
-
 export { default as
   Pagination,
 } from './packages/pagination';