+++ packages/icon/src/Icons.spec.js
@@ -5,7 +5,7 @@ import Icon from './Icon';
 describe('Icon', () => {
     it('should apply the correct icon-name', () => {
         const icon = mount(<Icon name="danger" />);
-        expect(icon.getDOMNode().className).toContain("fa-danger");
+        expect(icon.getDOMNode().className).toContain("ai-danger");
     });
 
     it('should apply the correct className', () => {
@@ -15,6 +15,6 @@ describe('Icon', () => {
 
     it('should set the data-qa attribute', () => {
         const icon = shallow(<Icon name="test" qa="id-1234" />);
-        expect(icon.find('.fa').props()).toHaveProperty('data-qa', 'id-1234');
+        expect(icon.find('.ai').props()).toHaveProperty('data-qa', 'id-1234');
     });
 });